Relocate cluster create command
